% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voi_remote.R
\name{evppi.remote}
\alias{evppi.remote}
\title{Calculate the expected value of partial perfect information from a decision model}
\usage{
evppi.remote(
  outputs,
  inputs,
  pars = NULL,
  se = FALSE,
  B = 500,
  nsim = NULL,
  verbose = FALSE,
  method = NULL,
  ...
)
}
\arguments{
\item{outputs}{This could take one of two forms

"net benefit" form: a matrix or data frame of samples from the uncertainty
distribution of the expected net benefit.  The number of rows should equal
the number of samples, and the number of columns should equal the number
of decision options.

"cost-effectiveness analysis" form: a list with the following named
components:

\code{"c"}: a matrix or data frame of samples from the distribution of
costs.  There should be one column for each decision option.

\code{"e"}: a matrix or data frame of samples from the distribution of
effects, likewise.

\code{"k"}: a vector of willingness-to-pay values.

Objects of class \code{"bcea"}, as created by the \pkg{BCEA} package, are in
this "cost-effectiveness analysis" format, therefore they may be supplied as
the \code{outputs} argument.

If \code{outputs} is a matrix or data frame it is assumed to be of "net
benefit" form.  Otherwise if it is a list, it is assumed to be of "cost
effectiveness analysis" form.}

\item{inputs}{Matrix or data frame of samples from the uncertainty
distribution of the input parameters of the decision model.   The number
of columns should equal the number of parameters, and the columns should
be named.    This should have the same number of rows as there are samples
in \code{outputs}, and each row of the samples in \code{outputs} should
give the model output evaluated at the corresponding parameters.}

\item{pars}{A character vector giving the parameters of interest, for which
a single EVPPI calculation is required.  If the vector has multiple
element, then the joint expected value of perfect information on all these
parameters together is calculated.

Alternatively, \code{pars} may be a list.  Multiple EVPPI calculations are
then performed, one for each component of \code{pars} defined in the above
vector form.

\code{pars} must be specified if \code{inputs} is a matrix or data frame.
This should then correspond to particular columns of \code{inputs}.    If
\code{inputs} is a vector, this is assumed to define the single parameter
of interest, then \code{pars} is not required.}

\item{se}{If possible, calculate a standard error for the EVPPI.  Currently
only supported for \code{method="gam"}.}

\item{B}{Number of parameter replicates for calculating the standard error.}

\item{nsim}{Number of simulations from the model to use for calculating
EVPPI.  The first \code{nsim} rows of the objects in \code{inputs} and
\code{outputs} are used.}

\item{verbose}{If \code{TRUE}, then messages are printed describing each step of
the calculation, if the method supplies these.  Useful to see the progress
of slow calculations.}

\item{method}{Character string indicating the calculation method. The default
methods are based on nonparametric regression}

\item{...}{Other arguments to control specific methods.}
}
\value{
A data frame with a column \code{pars} indicating the parameter(s)
and a column \code{evppi} giving the corresponding EVPPI.

If \code{outputs} is of "cost-effectiveness analysis" form so that there is
one EVPPI per willingness-to-pay value, then a column \code{k} identifies the
willingness-to-pay.

If standard errors are requested, then the standard errors are returned in
the column \code{se}.
}
\description{
Calculate the expected value of partial perfect information from a decision model
}
\examples{
set.seed(1)
nsam <- 10000
inputs <- data.frame(
  p1 = rnorm(nsam, 1, 1),
  p2 = rnorm(nsam, 0, 2)
)
outputs_nb <- data.frame(
  t1 = 0,
  t2 = inputs$p1 - inputs$p2
)
evppi.remote(outputs_nb, inputs, pars="p1")
}
