% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voi_remote.R
\name{evppi_mc.remote}
\alias{evppi_mc.remote}
\title{Title}
\usage{
evppi_mc.remote(model_fn, par_fn, pars, nouter, ninner, wtp, mfargs, verbose)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
