% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voi_remote.R
\name{evppivar.remote}
\alias{evppivar.remote}
\title{Calculate the expected value of partial perfect information for an estimation
problem.  This computes the expected reduction in variance in some quantity
of interest with perfect information about a parameter or parameters of interest.}
\usage{
evppivar.remote(
  outputs,
  inputs,
  pars = NULL,
  method = NULL,
  nsim = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{outputs}{This could take one of two forms

"net benefit" form: a matrix or data frame of samples from the uncertainty
distribution of the expected net benefit.  The number of rows should equal
the number of samples, and the number of columns should equal the number
of decision options.

"cost-effectiveness analysis" form: a list with the following named
components:

\code{"c"}: a matrix or data frame of samples from the distribution of
costs.  There should be one column for each decision option.

\code{"e"}: a matrix or data frame of samples from the distribution of
effects, likewise.

\code{"k"}: a vector of willingness-to-pay values.

Objects of class \code{"bcea"}, as created by the \pkg{BCEA} package, are in
this "cost-effectiveness analysis" format, therefore they may be supplied as
the \code{outputs} argument.

If \code{outputs} is a matrix or data frame it is assumed to be of "net
benefit" form.  Otherwise if it is a list, it is assumed to be of "cost
effectiveness analysis" form.}

\item{inputs}{Matrix or data frame of samples from the uncertainty
distribution of the input parameters of the decision model.   The number
of columns should equal the number of parameters, and the columns should
be named.    This should have the same number of rows as there are samples
in \code{outputs}, and each row of the samples in \code{outputs} should
give the model output evaluated at the corresponding parameters.}

\item{pars}{A character vector giving the parameters of interest, for which
a single EVPPI calculation is required.  If the vector has multiple
element, then the joint expected value of perfect information on all these
parameters together is calculated.

Alternatively, \code{pars} may be a list.  Multiple EVPPI calculations are
then performed, one for each component of \code{pars} defined in the above
vector form.

\code{pars} must be specified if \code{inputs} is a matrix or data frame.
This should then correspond to particular columns of \code{inputs}.    If
\code{inputs} is a vector, this is assumed to define the single parameter
of interest, then \code{pars} is not required.}

\item{method}{Character string indicating the calculation method. The default
methods are based on nonparametric regression}

\item{nsim}{Number of simulations from the model to use for calculating
EVPPI.  The first \code{nsim} rows of the objects in \code{inputs} and
\code{outputs} are used.}

\item{verbose}{If \code{TRUE}, then messages are printed describing each step of
the calculation, if the method supplies these.  Useful to see the progress
of slow calculations.}

\item{...}{Other arguments to control specific methods.}
}
\value{
expected reduction in variance in some quantity of interest with
perfect information about a parameter or parameters of interest
}
\description{
Calculate the expected value of partial perfect information for an estimation
problem.  This computes the expected reduction in variance in some quantity
of interest with perfect information about a parameter or parameters of interest.
}
\examples{
p1 <- rbeta(10000, 5, 95)
beta <- rnorm(10000, 0.8, 0.4)
p2 <- plogis(qlogis(p1) + beta)
inputs <- data.frame(p1, beta)
evppivar.remote(p2, inputs, par="p1")
}
