% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voi_remote.R
\name{evsivar.remote}
\alias{evsivar.remote}
\title{Title}
\usage{
evsivar.remote(
  outputs,
  inputs,
  study,
  datagen_fn,
  pars,
  n,
  likelihood,
  analysis_model,
  analysis_options,
  decision_model,
  Q,
  npreg_method,
  nsim,
  verbose,
  method,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
